# Small performance improvement on large repos