--- a/plugins/git/plugin.zsh
+++ b/plugins/git/plugin.zsh
@@ -54,7 +54,7 @@ prompt_geometry_git_branch() {
 }
 
 prompt_geometry_git_status() {
-  if test -z "$(git status --porcelain --ignore-submodules)"; then
+  if test -z "$(git status --porcelain --ignore-submodules HEAD)"; then
     echo $GEOMETRY_GIT_CLEAN
   else
     echo $GEOMETRY_GIT_DIRTY