▲ /path/to/large-repo time git status --porcelain --ignore-submodules &> /dev/null
    git status --porcelain --ignore-submodules &> /dev/null  0.69s user 0.03s system 99% cpu 0.733 total

    ▲ /path/to/large-repo time git status --porcelain --ignore-submodules HEAD &> /dev/null
    git status --porcelain --ignore-submodules HEAD &> /dev/null  0.20s user 0.00s system 99% cpu 0.206 total

It's a quite noticeable improvement but I don't know if I'm missing/breaking something up.

Partially fixes #60